+++ tests/test_interactive_apps.py
@@ -23,6 +23,7 @@
 #
 
 import unittest.mock as mock
+import argparse
 
 import pytest
 
@@ -38,14 +39,6 @@ def get_itm(tms):
     itm.onecmd_plus_hooks(args)
     return itm
 
-def parse_apps(lines):
-    apps = []
-    for line in lines.splitlines():
-        app = tm.models.TomcatApplication()
-        app.parse(line)
-        apps.append(app)
-    return apps
-
 ###
 #
 # list
@@ -56,18 +49,51 @@ def mock_apps(mocker):
     return mocker.patch('tomcatmanager.models.TomcatManagerResponse.result', create=True,
                          new_callable=mock.PropertyMock)
 
-def test_process_apps_empty():
+def parse_apps(lines):
+    """helper function to turn colon seperated text into objects"""
+    apps = []
+    for line in lines.splitlines():
+        app = tm.models.TomcatApplication()
+        app.parse(line)
+        apps.append(app)
+    return apps
+
+def test_list_process_apps_empty():
     lines = ''
     itm = tm.InteractiveTomcatManager()
-    args = itm.list_parse_args('')
+    args = itm._list_parse_args('')
     apps = parse_apps(lines)
-    grouped_apps = itm.process_apps(apps, args)
-    assert isinstance(grouped_apps, list)
-    assert len(grouped_apps) == 0
-    
-def test_list_raw(tomcat_manager_server, mock_apps, capsys):
-    raw_apps = """/:running:0:ROOT
-/shiny:stopped:17:shiny##v2.0.6
+    apps = itm._list_process_apps(apps, args)
+    assert isinstance(apps, list)
+    assert len(apps) == 0
+
+LIST_ARGV_BAD = [
+    '-raw',
+    '--raw -b tate',
+    '--by=version',
+    '-s loading',
+    '-sl',
+    '-bs',
+]    
+@pytest.mark.parametrize('argv', LIST_ARGV_BAD)
+def test_list_parse_args_failure(argv):
+    itm = tm.InteractiveTomcatManager()
+    with pytest.raises(SystemExit):
+        args = itm._list_parse_args(argv)
+    assert itm.exit_code == itm.exit_codes.usage
+
+@pytest.mark.parametrize('raw', ['', '-r', '--raw'])
+@pytest.mark.parametrize('state', ['', '-s running', '-s stopped', '--state=running', '--state=stopped'])
+@pytest.mark.parametrize('sort', ['', '-b state', '-b path', '--by=state', '--by=path'])
+def test_list_parse_args(raw, state, sort):
+    itm = tm.InteractiveTomcatManager()
+    argv = '{} {} {}'.format(raw, state, sort)
+    args = itm._list_parse_args(argv)
+    assert itm.exit_code == itm.exit_codes.success
+
+def test_list_sort_by_state(tomcat_manager_server, mock_apps, capsys):
+    raw_apps = """/shiny:stopped:0:shiny##v2.0.6
+/:running:0:ROOT
 /shiny:running:15:shiny##v2.0.7
 /host-manager:running:0:/usr/share/tomcat8-admin/host-manager
 /manager:running:0:/usr/share/tomcat8-admin/manager
@@ -76,11 +102,30 @@ def test_list_raw(tomcat_manager_server, mock_apps, capsys):
 /host-manager:running:0:/usr/share/tomcat8-admin/host-manager
 /manager:running:0:/usr/share/tomcat8-admin/manager
 /shiny:running:15:shiny##v2.0.7
-/shiny:stopped:17:shiny##v2.0.6
+/shiny:stopped:0:shiny##v2.0.6
 """ 
     mock_apps.return_value = raw_apps
     interactive_tomcat = get_itm(tomcat_manager_server)
-    interactive_tomcat.onecmd_plus_hooks('list --raw')
+    interactive_tomcat.onecmd_plus_hooks('list --raw -b state')
+    out, err = capsys.readouterr()
+    assert out == expected
+
+def test_list_sort_by_path(tomcat_manager_server, mock_apps, capsys):
+    raw_apps = """/:running:0:ROOT
+/shiny:stopped:0:shiny##v2.0.6
+/shiny:running:15:shiny##v2.0.7
+/host-manager:running:0:/usr/share/tomcat8-admin/host-manager
+/manager:running:0:/usr/share/tomcat8-admin/manager
+"""
+    expected = """/:running:0:ROOT
+/host-manager:running:0:/usr/share/tomcat8-admin/host-manager
+/manager:running:0:/usr/share/tomcat8-admin/manager
+/shiny:stopped:0:shiny##v2.0.6
+/shiny:running:15:shiny##v2.0.7
+""" 
+    mock_apps.return_value = raw_apps
+    interactive_tomcat = get_itm(tomcat_manager_server)
+    interactive_tomcat.onecmd_plus_hooks('list --raw -b path')
     out, err = capsys.readouterr()
     assert out == expected